Array_Prompts Abstractification
Performs necessary conversions to Array_Prompts infrastructure to remove hardcoding and allow it to be overridden by children on an individual CBM basis. Basically what this means is that we can quickly type up new prompts for each game and it should work with the current evaluator / grader architecture base.