LNI Fixed Prompts, Game Selection
Creates and implements the fixed prompts for Letter Name Identification for the manual in the evaluator. Creates a way to track the currently selected Game in DataManager and adds a script that allows scene transitions based on the currently selected game.

Note: Instructions screen is temporarily cut until we make a LNI instruction screen.