Validation Refactor
This commit introduces a validation middleman, Validation_Games, that requires GetValidInput() in addition to Validation_Parents Validator.

This essentially requires games to check both individual answers and if the whole scene is valid.

This allows us to re-use much of the logic implemented for Vocab for several other games, while being able to customize the validation of individual responses as they change to match the game.

Also includes some UI changes for LNI (non-random).